MenuReducers/Context now renamed to OrdersReducer/Context for exact description of its function. New MenuReducers/Context is created
